+++ src/main/java/com/pragmatic/dao/rowmapper/TransactionRowMapper.java
@@ -14,7 +14,7 @@ public class TransactionRowMapper implements RowMapper<Transaction> {
                 .transactionId(rs.getInt("transaction_id"))
                 .sourceAccountId(rs.getInt("source_account_id"))
                 .destinationAccountId(rs.getInt("destination_account_id"))
-                .amount(rs.getDouble("amount"))
+                .amount(rs.getBigDecimal("amount"))
                 .actionDate(rs.getDate("action_date"))
                 .build();
     }
