+++ src/main/java/com/pragmatic/model/Transaction.java
@@ -2,6 +2,7 @@ package com.pragmatic.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Date;
 
@@ -12,7 +13,7 @@ public class Transaction  {
     private Integer transactionId;
     private final Integer sourceAccountId;
     private final Integer destinationAccountId;
-    private final Double amount;
+    private final BigDecimal amount;
     @Builder.Default
     private Date actionDate = Date.from(Instant.now());
 }
